requirements.txt
agate==1.9.1
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
alembic==1.13.2
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
apache-airflow==2.10.0
apache-airflow-providers-celery==3.7.3
apache-airflow-providers-common-compat==1.1.0
apache-airflow-providers-common-io==1.4.0
apache-airflow-providers-common-sql==1.15.0
apache-airflow-providers-fab==1.2.2
apache-airflow-providers-ftp==3.10.1
apache-airflow-providers-http==4.12.0
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-sqlite==3.8.2
apispec==6.6.1
appnope==0.1.4
argcomplete==3.5.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
bcrypt==4.2.0
beautifulsoup4==4.12.2
billiard==4.2.0
blinker==1.8.2
bs4==0.0.1
cachelib==0.9.0
cachetools==5.3.2
celery==5.4.0
certifi==2023.7.22
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
colorama==0.4.6
colorlog==6.8.2
comm==0.2.2
ConfigUpdater==3.2
connexion==2.14.2
cramjam==2.8.3
cron-descriptor==1.4.3
croniter==3.0.3
cryptography==41.0.7
daff==1.3.46
db-dtypes==1.3.0
dbt-adapters==1.4.1
dbt-bigquery==1.8.2
dbt-common==1.7.0
dbt-core==1.8.6
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.5.1
debugpy==1.8.5
decorator==5.1.1
deepdiff==7.0.1
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.6.1
docker==7.1.0
docutils==0.16
EasyProcess==1.1
email_validator==2.2.0
entrypoint2==1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastparquet==2024.5.0
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flower==2.0.1
frozenlist==1.4.1
fsspec==2024.6.1
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.3.3
google-cloud-dataproc==5.11.0
google-cloud-documentai==2.31.0
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-re2==1.1.20240702
google-resumable-media==2.6.0
googleads==43.0.0
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.62.3
gspread==6.1.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
humanize==4.10.0
idna==3.4
importlib-metadata==6.11.0
importlib_resources==6.4.0
inflection==0.5.1
ipykernel==6.29.5
ipython==8.18.1
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kombu==5.4.0
lazop-sdk==1.0.2
lazy-object-proxy==1.10.0
leather==0.4.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
Logbook==1.5.3
lxml==5.3.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mashumaro==3.13.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
methodtools==0.4.7
minimal-snowplow-tracker==0.0.2
more-itertools==10.4.0
MouseInfo==0.1.3
msgpack==1.0.8
mss==9.0.1
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.8.1.78
openpyxl==3.1.2
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.1.3
paramiko==3.4.1
parsedatetime==2.6
parso==0.8.4
pathspec==0.12.1
pendulum==3.0.0
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.2.2
pluggy==1.5.0
prison==0.2.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.0
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyGetWindow==0.0.9
Pygments==2.18.0
pygsheets==2.0.6
PyJWT==2.9.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyobjc-core==10.0
pyobjc-framework-Cocoa==10.0
pyobjc-framework-Quartz==10.0
pyodbc==5.1.0
pyparsing==3.1.1
PyPDF2==3.0.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-nvd3==0.16.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytweening==1.0.7
pytz==2023.3.post1
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.0.8
referencing==0.35.1
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.1
rich-argparse==1.5.2
rpds-py==0.20.0
rsa==4.9
rubicon-objc==0.4.7
selenium==4.25.0
setproctitle==1.3.3
six==1.16.0
slackclient==2.9.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.53
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.1
sshtunnel==0.4.0
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.15.0
tornado==6.4.1
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2023.3
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.2
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==2.2.3
wget==3.2
wirerope==0.4.7
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
xlrd==2.0.1
xmltodict==0.13.0
yarl==1.9.4
zeep==4.2.1
zipp==3.20.0
